Remove custom placement handling
